chore: move test fixtures
